Improve roster sp (#25)
* Add multi threading on the resolutions of the subproblem. Disable by default. Add a ThreadsPool class in the tools to help with multi threading.

* Commented boost rc_spp. removed most of boost objects in the rc_spp.

* Add more comments about parallelization of subproblems.

* For disjoint strategy, only forbid shift with a positive dual cost.

* Fixed bug in the pricing of the short rotations. Because of the feasibility variables,  the bestCost could be to small as a first value (all the short succession have a higher cost). Also changed the settings for restart the strategy of the sublem to the default strategy: always restart when master is infeasible (needs a lot of new columns).

* Refactoring to separate the generic subproblem from the each implementation. Also refactoring of the RCGraph to separate the graph from the algorithm to solve it.

+++ src/data/Nurse.cpp
@@ -292,7 +292,7 @@ bool Preferences::wantsTheDayOff(int nurseId, int day) const{
 	if(itM == wishesOff_.at(nurseId).end())
 		return false;
   // Set does not repeat its elements. Wants the day off if and only if all shifts off (-1 because REST does not appear)
-	return itM->second.size() == nbShifts_-1;
+	return (int) itM->second.size() == nbShifts_-1;
 }
 
 // Total number of shifts off that the nurse wants
@@ -311,7 +311,7 @@ int Preferences::howManyDaysOff(int nurseId, int dayMin, int dayMax) const {
 	int nbDayOff = 0;
 	// look at every wishes of the nurse
 	for( const auto & dayOff: wishesOff_.at(nurseId)){
-		nbDayOff += ( (dayOff.first >= dayMin) && (dayOff.first <= dayMax) && (dayOff.second.size() == nbShifts_-1));
+		nbDayOff += ( (dayOff.first >= dayMin) && (dayOff.first <= dayMax) && ((int)dayOff.second.size() == nbShifts_-1));
 	}
 	return nbDayOff;
 }
@@ -355,7 +355,7 @@ bool Preferences::wantsTheDayOn(int nurseId, int day) const{
   if(itM == wishesOn_.at(nurseId).end())
     return false;
   // Set does not repeat its elements. Wants the day on if and only if all shifts off (-1 because REST does not appear)
-  return itM->second.size() == nbShifts_-1;
+  return (int)itM->second.size() == nbShifts_-1;
 }
 
 // Total number of shifts on that the nurse wants
@@ -374,7 +374,7 @@ int Preferences::howManyDaysOn(int nurseId, int dayMin, int dayMax) const {
   int nbDayOn = 0;
   // look at every wishes of the nurse
   for( const auto & dayOn: wishesOn_.at(nurseId)){
-    nbDayOn += ( (dayOn.first >= dayMin) && (dayOn.first <= dayMax) && (dayOn.second.size() == nbShifts_-1));
+    nbDayOn += ( (dayOn.first >= dayMin) && (dayOn.first <= dayMax) && ((int)dayOn.second.size() == nbShifts_-1));
   }
   return nbDayOn;
 }
